requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
httpx>=0.27.0
PyYAML>=6.0.2
python-dateutil>=2.9.0
rich>=13.9.0
typer>=0.12.0
